fix: remove old earnings report functions
